Move `wasmtime_cranelift::gc` to `wasmtime_cranelift::func_environ::gc`
Purely mechanical.

The `gc` module is the GC-related parts of the function environment, so it makes
sense to be a submodule of `func_environ`. Additionally, there are a few fields
in `FuncEnvironment` that are `pub(crate)` just so that they can be accessed
from the `gc` module, and after this commit, that will no longer be
necessary. This commit doesn't change that visibility, but follow ups will.