asm: improve DSL's understanding of prefixes (#10475)
* asm: improve DSL's understanding of prefixes

Previously, Cranelift's x64 backend only understood a subset of the
available prefixes that can be encoded and the documentation for these
was quite sparse. This change upgrades the assembler to understand all
kinds of prefixes, even those as yet unused, and drastically increases
the amount of documentation around these prefixes. This, along with some
new `Group`-based typing, should prevent accidental misuse but also
increase confidence that this implementation is correctly covering the
possible cases.

* review: check for empty byte slice