asm: fix trace logging again
As pointed out in #10476, there are a few more locations in which the
raw encoding of a register is used when pretty-printing memory operands.
Using the raw encoding before register allocation results in a panic.
This change ensures that the additional locations delegate to the
`to_string()` method which is overriden in `cranelift-codegen` to
know how to pretty-print virtual registers. Closes #10476.