asm: add `LOCK`-prefixed instructions (#10477)
* asm: add initial `LOCK`-prefixed instructions

This change is the first in a series to add new instructions that emit
the `LOCK` prefix during encoding.

* asm: add all `lock and*` instructions

This defines all `LOCK`-prefixed variants for `and*`.

* asm: add all `lock add|adc*` instructions

* asm: add all `lock or*` instructions

* asm: add all `lock sub|sbb*` instructions

* asm: add all `lock xor*` instructions

* asm: reformat using slightly longer lines

This allows the `LOCK`-prefixed definitions to still fit on a single
line.

* asm: integrate mem-only operands in `cranelift-codegen`

`LOCK`-prefixed instructions can only write to memory operands (see
prior commits). This change wires up the necessary codegen to handle the
new `m*` operands.

* asm: lower RMW instructions with new `LOCK`-prefixed instructions

* Bump allowed size of the `Inst` enum

It is unclear why this would happen now, though perhaps the addition of
all the `LOCK`-prefixed instructions could do it (?).