Improve errors with pooling-by-default `serve` command (#10483)
* Improve errors with pooling-by-default `serve` command

This commit is intended to address #10482 where the defaults of the
`wasmtime serve` subcommand produced a confusing and surprising error.
Specifically the `-Wmax-table-elements` option, prior to this change, only
affected the store limiter used and didn't actually affect the pooling
allocator settings. That meant that if a module exceeded the limits of
the pooling allocator it would produce an error message that seemed like
`-Wmax-table-elements` would fix but it wouldn't actually.

Two changes in this commit are meant to address this:

* Errors from the pooling allocator have been update to mention "pooling
  allocator" within them somewhere to surface where the error is coming
  from.
* The `-Wmax-memory-size` and `-Wmax-table-elements` configuration are
  now applied to the pooling allocator automatically if the
  corresponding `-Opooling-*` option isn't passed.

That should mean that the original error should be a bit easier to debug
while the attempted solution will also work.

* Fix test expectations

* Get new test passing on 32-bit