Add missing `VMContext` field for component table use
When components access tables, these may lazily initialized. Eventually,
we will need to call `Instance::get_defined_table_with_lazy_init` to set
up the table immediately prior to use. To do so, this change stores an
additional `*mut VMContext` (for a core instance, not a component) in
the `VMComponentContext` along with each `*mut VMTableDefinition` so
that we can recover the core instance needed for lazy initialization.