Use `VMTableImport` to record component tables
To avoid extra structures, this change uses `VMTableImport` everywhere a
component would need to keep track of a `VMTableDefinition` and
`VMContext` (recall the `VMContext` is needed for lazy initialization).
This is a bit awkward since this usage is not truly an import, but
perhaps we can change the name later.