Improve table support for components (#10499)
* Fix documentation, add debug check

This change cleans up some missing items from #10497: a copy-paste typo
and some missing debug assertions about the initial state of the table
list in the `VMComponentContext`.

* Add missing `VMContext` field for component table use

When components access tables, these may lazily initialized. Eventually,
we will need to call `Instance::get_defined_table_with_lazy_init` to set
up the table immediately prior to use. To do so, this change stores an
additional `*mut VMContext` (for a core instance, not a component) in
the `VMComponentContext` along with each `*mut VMTableDefinition` so
that we can recover the core instance needed for lazy initialization.

* Use `VMTableImport` to record component tables

To avoid extra structures, this change uses `VMTableImport` everywhere a
component would need to keep track of a `VMTableDefinition` and
`VMContext` (recall the `VMContext` is needed for lazy initialization).
This is a bit awkward since this usage is not truly an import, but
perhaps we can change the name later.