Cranelift: update to regalloc2 v0.11.2.
This incorporates a fix for a panic caused by too many live registers
that is not properly returned as a regalloc error. This condition should
not be reachable from Cranelift (which should not generate constraints
that trigger this condition), and was seen only during development where
different constraints were added; but this PR is an update out of an
abundance of caution.