Plumb initial translation of `thread.spawn_indirect`
This change in no way implements actual spawning of threads; it simply
translates the CM builtin `thread.spawn_indirect` to invoke a real
Wasmtime libcall that does some preliminary checks and panics. It is
useful, though, to (1) exercise all the new table-related functions and
(2) show the limits of Wasmtime's current state: we will need a way of
interacting with shared functions and possibly some form of shared
store.