Run fewer wast tests in CI (#10536)
Currently the `wast` test suite does a combinatorial matrix of `compiler
x pooling x allocator` to produce the list of tests to run. We also have
a `wast_tests` fuzzer, however, which configures even more knobs and run
tests under these configurations. In the interest of saving time on CI
and avoiding too much combinatorial explosion this commit changes the
`wast` test suite to avoiding combinatorics and instead running in only
"interesting" combinations.

The goal here is to weed out common issues on CI quickly, such as
compiler-specific behaviors, but leave the full spectrum of
combinatorial tests to fuzzing. For example we probably aren't getting
much extra coverage running Winch with the pooling allocator over
running Cranelift with the pooling allocator.

This reduces the number of tests from ~4500 to ~2500 which should
hopefully lead to some minor improvements in CI time.