s390x: Skip memory allocation failure tests (#10530)
These tests assume that a very large allocation will fail due to
exceeding address space limits.  This is not the case on s390x,
where the full 64-bit address space is available for allocation.
(Of course, actually using that much memory would still fail at
some point, but the mere allocation may not, depending on the
kernel's memory overcommit handling configuration.)

Note that running the tests under qemu will cause the expected
failure, which is why this problem is not noticed by the CI.

Simply disable these test on s390x.