Cranelift: riscv64: fix instruction worst-case-length checks.
In #10502, we introduced changes that could make callsites be
arbitrarily long, because they now include loads of
return-values-on-stack. We made use of the existing island mechanism
(now presented as a new pseudoinst as in aarch64, rather than as ad-hoc
emission code) to ensure that we meet label-reference-distance
deadlines. Unfortunately we didn't update the debug-assert that checks
instructions for worst-case size to exclude calls (and the new
`EmitIsland` pseudoinst), since they handle islanding separately.

Found via fuzzbug at [1].

[1]: https://oss-fuzz.com/testcase-detail/4819793142415360