cranelift-jit: add bump memory allocator, make configurable (#10512)
* cranelift-jit: add bump memory allocator, make configurable

This adds a new JIT memory allocator which operates on an address range
that is reserved up-front. This is useful in order to side-step issues
where allocation behavior affect relocations, i.e. on x86 where we run
into issues when the system allocator returns memory that is more than
4 GiB apart.

The main drawback of this allocator option is that the memory range has
to be reserved up-front, so allocation will fail if this space is
exhausted.

This commit:
- adds a 'JITMemoryProvider' trait
- moves the previous on-demand allocator to this trait
- adds a new 'Arena' allocator

* ArenaMemoryProvider: add tests, simplify alignment logic

* ArenaMemoryProvider: add asserts

* Move pipeline flush back out of `set_readable_and_executable`

* Fix aarch64 build

* Document Windows overcommit limitation