threads: add feature flags (#10206)
* threads: add feature flags

This adds both the Cargo-level and CLI-level flags for the
shared-everything-threads proposal.

* Remove Cargo-level feature flags

As recommended in a review, we can probably use the `threads` feature
flag instead for the same kind of conditional compilation.

* Remove CLI-level flags

Since we don't expect users to be interacting with
shared-everything-threads modules from the command line, hide this for
now.

* Add internal flags necessary for testing