Don't discard errors in `default_value` helpers (#10584)
This commit updates the `default_value` helpers added recently to return
a `Result` instead of returning an `Option<T>` and throwing away error
information. This fixes a fuzz bug showing up recently which happened
because the error in question was one we've flagged to ignore, but
because the error was discarded we didn't know to ignore it so it ended
up causing a fuzz failure.