Canonicalize NaN in cranelift-fuzzgen
Lacking this has happened to work fine up to now but Rust nightly is
changing float implementations which is changing float payloads. As a
result this is the only reliable method of ensuring that differential
fuzzing between the interpreter and native backends remains reliable, in
the same manner as this being enabled for core wasm as well.

Closes #10583