Fix `FreeList::add_capacity` when we still don't have enough capacity for first allocation (#10589)
When we haven't added enough capacity such that it is greater than or equal to
our alignment, we were triggering a debug assertion and would get integer
overflow and a corrupted free list in release builds. Note that, due to our GC
heap sandboxing, this couldn't lead to anything worse than a panic/abort.