Fix the new `logged-filetests.rs` test (#10605)
This test unfortunately didn't actually have any effect in CI. Locally
setting `RUST_LOG` and running `cargo run test filetests` showed a panic
in the s390x backend that this test was not uncovering. The reason that
this test wasn't actually doing anything on CI is that it's tested
adjacently with the `wasmtime-cli` crate. Both crates are tested with
`--all-features` which enables the CLI's `disable-logging` feature which
meant that all logging statements, including the ones we want to test,
were compiled out.

This commit fixes the s390x panic and then additionally adds a special
CI job (repurposing an existing one) for testing just this test in
isolation.