component::__internal::InstanceType now owns its contents
Having refs to a pair of arcs made it less expensive to create an
InstanceType, but tied its to a borrow of the ComponentInstance
(transitively the Store) which makes some new typechecking operations
which have an AsContextMut impossible.

This has no direct benefit in this PR, but it split out of #10610
because it is more easily reviewed in isolation.