Add hidden Instance::instance_pre method
which retrieves an InstancePre corresponding to an Instance.

All of the information needed to compute the InstancePre is already
present in the Store under the Instances's index.

This method is pub but doc(hidden) because it is just for typechecking
(to be used in the bindings generator), so like other accessors of
InstanceType or typecheck methods, it is not exposed in the public API.