wasmtime component: get InstancePre from Instance (#10621)
* component::InstancePre contains the types::Component with imported resource types

types::Component can be used to create an InstanceType (also added as a
hidden method). The construction of InstancePre in the Linker calculates
this information already, and the types::Compoonent is just storing Arcs
to this information.

* Add hidden Instance::instance_pre method

which retrieves an InstancePre corresponding to an Instance.

All of the information needed to compute the InstancePre is already
present in the Store under the Instances's index.

This method is pub but doc(hidden) because it is just for typechecking
(to be used in the bindings generator), so like other accessors of
InstanceType or typecheck methods, it is not exposed in the public API.

* InstancePre: rather than keep a ComponentType, just keep the resource type table

construct InstanceType right from the contents of InstancePre instead of
using the intermediate ComponentType. The net is that InstancePre is
just one word larger & one arc clone more expensive than before.