asm: change how regalloc interacts with operands
This PR changes the paradigm for how a user (like `regalloc2`) will
interact with the operands of an assembler instruction. Instead of using
a visitor pattern, calling functions for each of the operands, this
approach builds up a list of operands that expose `&mut` references to
the underlying operands of the instruction.

While this does have the advantage of being a bit more straightforward,
the motivation for this is more subtle: some instructions have special
register constraints that the visitor pattern does not handle. E.g.,
when switching `AluConstOp` to use the new assembler, I noticed that
the calls it makes to the register allocator are quite different than
the visitor expects: `AluConstOp` boils down to `xor %r*, %r*` and
expects us a completely different call to `regalloc2`: a single
`reg_def()`. We would need to build a completely different visitor
implementation for this.

Also, this refactoring makes it possible for users of the assembler to
examine and modify any operands to any instruction. This should be
useful regardless of the `AluConstOp` motivation.