asm: plumb fixed registers through (#10629)
* asm: plumb fixed registers through

This change introduces a new `Fixed` wrapper type that allows us to
plumb fixed registers through the assembler. These instructions are not
yet used by `cranelift-codegen`, but will be necessary at some point as
described in [#10238]. The end result of this is that up at the
`cranelift-codegen` level, fixed registers should look like
run-of-the-mill virtual registers using the appropriate types, but down
in `cranelift-assembler-x64` we panic if the register allocator does not
give them the correct register.

[#10238]: https://github.com/bytecodealliance/wasmtime/issues/10238

* fix: reduce `size_of::<Inst>` test

It is unclear why adding more `Fixed<R>` fields would somehow _reduce_
the size of `Inst`. What is quite clear is that I do not understand how
Rust chooses to lay out these fields.