Remove `AluRmiR`
This change completely removes the `Inst::AluRmiR` variant along with
associated code. The only remaining place this was used was via the
`Inst::alu_rmi_r` helper, which is replaced by `Inst::add|sub|and|or...`
helpers.

The lowering, implemented here in Rust, is meant to mimic the behavior
also-encoded in ISLE. In discussions about this in the Cranelift
meeting, this duplication was of low concern. Nevertheless, this could
be improved in the future. In any case, this commit clearly identifies
the locations that build an ALU instruction in Rust: ABI logic, Winch,
and long-ish lowering sequences.