asm: simplify visiting `Amode` registers (#10660)
Previously, the `Amode::registers_mut` method returned a list of
registers that generated code iterated over. This was because it was
difficult to capture the type needed for constraining the `Amode` (i.e.,
`Registers::ReadGpr`). This change moves that functionality to a helper
function with identical behavior, simplifying the generated code and API
surface.