Support custom yield behavior for epoch interrupts
As described in
https://github.com/bytecodealliance/wasmtime/issues/10667,
the yield future provided in-tree when a epoch callback cannot always
trigger the specific behavior desired in async schedulers like tokio.

This change introduces a new variant to the UpdateDeadline callback
with the future implementing the desired task suspension behavior
desired for yielding control within the async runtime.

Note that this change does not allow for controlling this behavior
for fuel yielding at this time.