Consolidate building guest programs for tests (#10654)
* Consolidate building guest programs for tests

This commit folds all C/C++ tests for debuginfo into the same location
as Rust programs compiled to wasm. That means that all of
`tests/all/debug/testsuite/*` is moved to
`crates/test-programs/src/bin/*`. Additionally the hardcoded per-test
configuration in `crates/test-programs/artifacts/build.rs` was removed
in favor of in-test configuration at the top of the file. This means
that all precompiled binaries are removed from the tree for DWARF
testing.

Various offsets were updated, a test was dropped as it wasn't used, and
some commands/directives were updated to account for this adjustment.

prtest:debug

* Adjust a lint setting