Simplify module features in serialization (#10672)
* Simplify module features in serialization

This commit simplifies how wasm features are handled in module
serialization. Instead of serializing a struct-of-bools this uses the
bitflags representation that wasmparser has. There are prior checks to
ensure that the wasmtime version for a serialized artifact is matching
so we know that the bits mean the same things too.

This reduces the boilerplate in implementing new wasm proposals and
should hopefully make it so that we get all the same benefit as before
just with less maintenance over time.

* Review comments

* Fix tests