[DWARF] Fix "simulated" DWARF (#10681)
* Add a test

Relying on the implementation detail of LLVM
not adding DWARF for module-level inline asm
is a bit ugly, but it is simpler than twisting
the "build system" we now have for C to build
an intermediate object without "-g".

* Fix simulated DWARF

We need to record the correct ranges for the CU and include
an offset zero entry in the line table, otherwise LLDB won't
class our functions as "having debug info".