Convert Winch to use assembler ALU instructions
Up until now, Winch had been using the `Inst::alu_rmi_r` for generating
the instructions it needed to emit. With this helper now gone, we have a
chose: (a) continue using `Inst` helpers that use the new assembler or
(b) use the new assembler directly. This change adopts option (b).

Because of how `cranelift-codegen`'s register allocation works, we have
been passing a `PairedGpr` into the read-write slots of the new
assembler (this enables us to separate out the read-side and write-side
as separate virtual registers). For Winch to use this, we have to expose
`PairedGpr` via the now-public `isa::x64::inst::external` module.