Make a non-deterministic log message deterministic
When generating a file we currently log the full path to that file which
changes depending on where the build is run and can prevent cache hits
in build systems like Bazel.

This change instead just prints the file name, along with a
deterministic name for the parent directory (e.g. `OUT_DIR` or
`ISLE_DIR`) which can be makes the log message deterministic and should
still be useful for debugging.