Fix up Cranelift emit tests
In the `Inst::<alu>` helpers, I chose to use the `rm` format of the ALU
instructions and not the previously-used `mr` format. Either format is
fine for reg-reg operations but but would be incorrect if any previous
uses of these helpers had inserted the `LOCK` prefix (`0xF0`)
immediately prior; the `LOCK` prefix is only valid on instructions with
a destination memory operand (i.e., `mr`, where `m` is read-write). I
did not find any uses like this and this would have been quite an unsafe
way to encode instructions anyways. I chose to use `rm` because it fits
the helper signature better: if we ever decided to use the memory
variant of the read operand (i.e., `src: RegMemImm`), then the `rm`
format is the right thing to use. But this change of format means that
we emit slightly different instructions for these long, atomic
sequences.