Collapse resource tables to be per-instance, not per-resource
This commit is an update to how Wasmtime represents and processes
resource handles for components. The upstream specification changed in
WebAssembly/component-model#427 and while it's been awhile since that
change this is Wasmtime catching up to that change. It's expected that
these paths are going to get stressed more with component model async
and more kinds of handles so this additionally refactors things to leave
sort of "holes" where async resources are going to go.

This notably will result in different behavior for guest programs.
Previously if there were two resources imported into a guest they could
have overlapping indices and now they're going to all have disjoint
indices since they're all in the same index space. It's expected that
this is will have a minimal, if any, impact on component guests as in
theory they're all mostly treating handles as opaque indexes today
anyway.