Update stability docs for the Wasm GC implementation (#10705)
Updates the "finished" and "fuzzed" columns to reflect their partial completion.

Removes the footnote about regressing non-GC-using Wasm performance when
enabled, as that is not actually true. We only do subtype checks as a fallback
when direct equality fails, and for non-GC-using Wasm if direct equality fails
then a trap is raised and the program terminates so a subtype check on this
branch doesn't matter. When direct equality succeeds, then there is no overhead
compared to when the GC proposal is not enabled.