x64: remove `Inst::AluRmiR` (#10687)
* Remove `AluRmiR`

This change completely removes the `Inst::AluRmiR` variant along with
associated code. The only remaining place this was used was via the
`Inst::alu_rmi_r` helper, which is replaced by `Inst::add|sub|and|or...`
helpers.

The lowering, implemented here in Rust, is meant to mimic the behavior
also-encoded in ISLE. In discussions about this in the Cranelift
meeting, this duplication was of low concern. Nevertheless, this could
be improved in the future. In any case, this commit clearly identifies
the locations that build an ALU instruction in Rust: ABI logic, Winch,
and long-ish lowering sequences.

* Make `Inst::<alu>` emit smaller immediate code

To avoid many offset changes in disassembly tests, this change improves
the `Inst::<alu>` helpers by teaching them how to emit smaller `mi`
encodings when the immediate part can fit in an 8-bit slot.

* Bless all affected Cranelift tests

* Fix up Cranelift emit tests

In the `Inst::<alu>` helpers, I chose to use the `rm` format of the ALU
instructions and not the previously-used `mr` format. Either format is
fine for reg-reg operations but but would be incorrect if any previous
uses of these helpers had inserted the `LOCK` prefix (`0xF0`)
immediately prior; the `LOCK` prefix is only valid on instructions with
a destination memory operand (i.e., `mr`, where `m` is read-write). I
did not find any uses like this and this would have been quite an unsafe
way to encode instructions anyways. I chose to use `rm` because it fits
the helper signature better: if we ever decided to use the memory
variant of the read operand (i.e., `src: RegMemImm`), then the `rm`
format is the right thing to use. But this change of format means that
we emit slightly different instructions for these long, atomic
sequences.

* Convert Winch to use assembler ALU instructions

Up until now, Winch had been using the `Inst::alu_rmi_r` for generating
the instructions it needed to emit. With this helper now gone, we have a
chose: (a) continue using `Inst` helpers that use the new assembler or
(b) use the new assembler directly. This change adopts option (b).

Because of how `cranelift-codegen`'s register allocation works, we have
been passing a `PairedGpr` into the read-write slots of the new
assembler (this enables us to separate out the read-side and write-side
as separate virtual registers). For Winch to use this, we have to expose
`PairedGpr` via the now-public `isa::x64::inst::external` module.

* Use `cranelift-assembler-x64` unconditionally in Winch

Due to how Cargo optional dependencies work and how we have chosen to
implement backend Cargo features, we decided in the Cranelift meeting to
just include this dependency unconditionally.

* Bless all affected Wasmtime tests

* review: refactor to remove the need for `Inst::<alu>` helpers entirely

* review: add `Inst::{add|sub}q_mi` helpers

This reduces the number of places we need to check for the opportunity
for smaller encodings.

* review: accept `impl Into<...>` in assembly constructors

This further reduces the boilerplate involved in creating new
instructions.